// src/data/openapi/models/searchFilterRangeType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * marketi
 * OpenAPI spec version: 1
 */

export type SearchFilterRangeType =
  (typeof SearchFilterRangeType)[keyof typeof SearchFilterRangeType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SearchFilterRangeType = {
  range: "range",
} as const;