// src/data/openapi/models/searchFilterSelectableType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * marketi
 * OpenAPI spec version: 1
 */

export type SearchFilterSelectableType =
  (typeof SearchFilterSelectableType)[keyof typeof SearchFilterSelectableType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SearchFilterSelectableType = {
  checkbox: "checkbox",
  toggle: "toggle",
} as const;