// src/data/openapi/models/searchFilterRange.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * marketi
 * OpenAPI spec version: 1
 */
import type { SearchFilterRangeType } from "./searchFilterRangeType";

export interface SearchFilterRange {
  type: SearchFilterRangeType;
  id: string;
  label: string;
  min: number;
  max: number;
  selectedMin: number;
  selectedMax: number;
}