// src/data/openapi/models/searchFilterSelectable.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * marketi
 * OpenAPI spec version: 1
 */
import type { SearchFilterSelectableType } from "./searchFilterSelectableType";
import type { SearchFilterVariant } from "./searchFilterVariant";

export interface SearchFilterSelectable {
  type: SearchFilterSelectableType;
  id: string;
  label: string;
  variants: SearchFilterVariant[];
}